requirements.txt
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-xlib==0.33
six==1.17.0
